requirements.txt
guessit == 3.8.0
fastapi == 0.104.1
uvicorn == 0.23.2
requests == 2.32.4
psycopg2 == 2.9.10
python-dotenv == 1.1.1
simple-log-factory == 0.0.1
openai == 1.98.0